wip: v7-clean-architecture in-flight changes
Snapshot of in-flight v7 work prior to applying the v-major release-fix
plan (docs/superpowers/plans/2026-05-16-v-major-release-fixes.md).
Captures producer / consumer / aggregator / persistence / docs / tests
changes that were sitting uncommitted in the working tree, plus the
code-review notes and CI workflow scaffolding that landed alongside.

+++ src/ServiceConnect.Interfaces/Persistence/IVersioned.cs
@@ -5,10 +5,15 @@ namespace ServiceConnect.Interfaces;
 /// concurrency control. Lets persistence callers read the version without dynamic
 /// dispatch or reflection against the wrapping type.
 /// </summary>
+/// <remarks>
+/// Version is <see cref="long"/>: <see cref="int"/> overflows after ~2.1B updates,
+/// which is unreachable for any realistic saga, but the typing change is free
+/// (matches MongoDB BSON Int64 natively) and forecloses the failure mode entirely.
+/// </remarks>
 public interface IVersioned
 {
     /// <summary>
     /// Gets the persistence version used for optimistic concurrency control.
     /// </summary>
-    int Version { get; }
+    long Version { get; }
 }
